+++ com/blazingkin/interpreter/executor/tensor/GetTensorValue.java
@@ -20,7 +20,7 @@ public class GetTensorValue implements InstructionExecutor {
 		BLZTensor<Integer> tensor = (BLZTensor<Integer>) tensorHuh.value;
 		int[] indices = new int[args.length-2];
 		for (int i = 1; i < args.length-1; i++){
-			indices[i-1] = Variable.getIntValue(Variable.getValue(args[i]));
+			indices[i-1] = Variable.getIntValue(Variable.getValue(args[i])).intValue();
 		}
 		Variable.setValue(args[args.length-1], new Value(VariableTypes.Integer,tensor.getTensorEntry(indices)));
 	}
