+++ com/blazingkin/interpreter/executor/timing/Wait.java
@@ -13,7 +13,7 @@ public class Wait implements InstructionExecutor {
 	public void run(String args[]){
 		Value val = Variable.getValue(args[0]);
 		if (Variable.isDecimalValue(val)){
-			double time = Variable.getDoubleVal(val);
+			double time = Variable.getDoubleVal(val).doubleValue();
 			try {
 				Thread.sleep((long)time);
 			} catch (InterruptedException e) {
