+++ com/blazingkin/interpreter/variables/VariableTypes.java
@@ -1,13 +1,16 @@
 package com.blazingkin.interpreter.variables;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 import com.blazingkin.interpreter.executor.lambda.LambdaExpression;
 
 @SuppressWarnings("rawtypes")
 public enum VariableTypes {
 
-	Integer(java.lang.Integer.class),
+	Integer(BigInteger.class),
 	String(java.lang.String.class),
-	Double(java.lang.Double.class),
+	Double(BigDecimal.class),
 	Boolean(java.lang.Boolean.class),
 	Rational(BLZRational.class),
 	Tensor(BLZTensor.class),
